requirements.txt
Django==5.2.4
Pillow==11.3.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-environ==0.12.0
whitenoise==6.9.0
black==25.1.0
flake8==7.3.0
pytest-django==4.11.1
coverage==7.3.2
djangorestframework==3.14.0
django-extensions==3.2.3
django-ckeditor==6.7.3
django-meta==2.2.0
django-anymail==10.2
redis==5.0.1
django-redis==5.4.0
django-security==0.12.0
sentry-sdk==1.38.0
django-taggit==5.0.1
django-mptt==0.14.0
django-haystack==3.2.1
django-import-export==3.3.1
django-allauth==65.9.0